refactor into a test suite runner
